one release note per FileBatch
